Create grasp keyword and adjacent sentence
Note: 'list' is not a very friendly class and it is not easy to manipulate. Change data frame is a easy way to deal with this kind of problem. 

Note: R is based on MATRIX. Try to think matrix manipulation before using loop. Save computing time and decrease the complexity.